Checking in WIP to save session leader names.
While working with session ids and proc.sname, I ran into cases where
proc.sname couldn't be found as the process with id <sid> had already
exited.

This commit is one way to solve this problem, by saving session leader
names in each process so it's not dependent on the session leader being
around. Another simpler way is to change the semantics of proc.sname to
walk the parent heirarchy to find the eldest process that still exists
and has the same session id and call it the session leader. That
approach will be in another branch.