WIP to add optional support for multiple filters.
Checking in WIP to add support for an additional set of individual
filters that can run alongside the main filter set in
sinsp::set_filter(). Each additional filter comes with a list of event
types for which the filter should run.

A multimap m_filter_by_evttype maps from event type to (check id,
sinsp_filter *). The individual filters are also held in a list
m_evttype_filters so they can be cleaned up on close().

A new method run_filters_on_evt() handles the details of running the
main filter as well as the individual filters. It finds the filters
related to the event's type, and runs those filters. If a filter
matches, the originally provided check_id is set via set_check_id().

My use case involves populating a filter object in lua using the lua
callbacks in lua_parser{_api}, so also modify lua_parser to add a
optional flag to get_parser that steals the m_filter object and creates
a new one.

TODO list:

 - add typedefs to avoid bare use of things like uint16_t for event
   types, uint32_t for check_ids, etc.
 - See if I can use auto and range loops more effectively.
 - I may want to also add negative event conditions (i.e. any
   event *not* in the provided list), or a wildcard, or similar.
 - Move the common reset code in lua_parser::get_filter duplicated from
   the contructor into its own function.