Switch to using an array of lists.
Switch the collection of (event type, filter) pairs to be held using an
array of lists instead of a multimap. The set of keys is
well-known (0..PPM_EVENT_MAX) so it's easy to just create an array of
list pointers, which are created/filled in on demand.

Also, we don't actually need to keep track of a check id for each filter
as the filter checks are still stamped with the check id. So drop that
argument from add_evttype_filter().