Switch back to erase/remove_if for sanitization.
It turns out that with -O3 (release flags), using erase/remove_if with
an operator struct is slightly faster than the inline version that was
here, as the compiler can inline the operators. It isn't faster with -O2
and is much slower without optimization.

Optimize for the release build by adding the operators back.